Refactor into layer-based architecture with configurable date pattern
- Replace monolithic copy_masters.py with 6 focused modules (scanning, path_utils, file_operations, flat_batch_selection, orchestration, statistics)
- Add EXIT_SUCCESS and EXIT_ERROR constants to config.py
- Make date directory pattern configurable via --date-dir-pattern CLI flag
- Remove duplicate default value definitions across function layers
- Use keyword-only parameters in process_blink_directory for better API clarity
- Update README with flexible flat matching documentation and new CLI flags
- Split test_copy_masters.py into module-specific test files
