+++ ap_copy_master_to_blink/__main__.py
@@ -7,14 +7,14 @@ Generated By: Claude Code (Claude Sonnet 4.5)
 import argparse
 import sys
 from pathlib import Path
-from typing import Dict, Optional, Tuple
+from typing import Optional, Tuple
 
 from ap_common import setup_logging, replace_env_vars, resolve_path
 
-from .copy_masters import process_blink_directory
-
-EXIT_SUCCESS = 0
-EXIT_ERROR = 1
+from . import config
+from .config import EXIT_SUCCESS, EXIT_ERROR
+from .orchestration import process_blink_directory
+from .statistics import print_summary, get_exit_code
 
 
 def validate_directories(
@@ -62,51 +62,6 @@ def print_header(library_dir: Path, blink_dir: Path, dry_run: bool) -> None:
     print(f"Library: {library_dir}")
     print(f"Blink:   {blink_dir}")
     print(f"Dry-run: {dry_run}")
-    print(f"{'='*70}\n")
-
-
-def print_summary(stats: Dict[str, int]) -> None:
-    """
-    Print summary of processing results.
-
-    Args:
-        stats: Statistics dictionary from process_blink_directory
-    """
-
-    def plural(count: int, singular: str) -> str:
-        """Format count with singular/plural form."""
-        return f"{count} {singular}{'s' if count != 1 else ''}"
-
-    def status_indicator(present: int, needed: int) -> str:
-        """Return status indicator based on present vs needed."""
-        return "ok" if present >= needed else "MISSING!"
-
-    print(f"\n{'='*70}")
-    print("Summary")
-    print(f"{'='*70}")
-    print(
-        f"Frames: {stats['frame_count']} lights "
-        f"({plural(stats['target_count'], 'target')}, "
-        f"{plural(stats['date_count'], 'date')}, "
-        f"{plural(stats['filter_count'], 'filter')})"
-    )
-
-    # CRITICAL: Output order MUST be: Biases, Darks, Flats
-    # CRITICAL: Bias MUST ALWAYS be shown (even if 0 of 0)
-    # This order has regressed multiple times - DO NOT CHANGE without updating tests
-    # See test_print_summary_output_order() in tests/test_copy_masters.py
-    print(
-        f"Biases: {stats['biases_present']} of {stats['biases_needed']} | "
-        f"{status_indicator(stats['biases_present'], stats['biases_needed'])}"
-    )
-    print(
-        f"Darks:  {stats['darks_present']} of {stats['darks_needed']} | "
-        f"{status_indicator(stats['darks_present'], stats['darks_needed'])}"
-    )
-    print(
-        f"Flats:  {stats['flats_present']} of {stats['flats_needed']} | "
-        f"{status_indicator(stats['flats_present'], stats['flats_needed'])}"
-    )
     print(f"{'='*70}\n")
 
 
@@ -191,6 +146,15 @@ Examples:
         "(default: 5). Only used with --flat-state.",
     )
 
+    parser.add_argument(
+        "--date-dir-pattern",
+        type=str,
+        default=config.DEFAULT_DATE_DIR_PATTERN,
+        metavar="PATTERN",
+        help="regex pattern to match date directory where masters are copied "
+        '(default: "^DATE_.*"). Assumes other tools create matching directories.',
+    )
+
     args = parser.parse_args()
 
     # Setup logging
@@ -225,24 +189,20 @@ Examples:
         scale_darks=args.scale_dark,
         flat_state_path=flat_state_path,
         picker_limit=args.picker_limit,
+        date_dir_pattern=args.date_dir_pattern,
     )
 
     # Print summary
     if not args.quiet:
         print_summary(stats)
 
-    # Return non-zero if any masters are missing
-    darks_missing = stats["darks_needed"] - stats["darks_present"]
-    flats_missing = stats["flats_needed"] - stats["flats_present"]
+    # Determine exit code from statistics
+    exit_code = get_exit_code(stats)
 
-    if darks_missing > 0 or flats_missing > 0:
-        if not args.quiet:
-            print(
-                "Warning: Some master frames are missing. Check logs above for details."
-            )
-        return EXIT_ERROR
+    if exit_code != EXIT_SUCCESS and not args.quiet:
+        print("Warning: Some master frames are missing. Check logs above for details.")
 
-    return EXIT_SUCCESS
+    return exit_code
 
 
 if __name__ == "__main__":
