+++ ap_copy_master_to_blink/file_operations.py
@@ -1,0 +1,86 @@
+"""
+File system operations for calibration frames
+
+Generated By: Claude Code (Claude Sonnet 4.5)
+"""
+
+from pathlib import Path
+from typing import Dict, Optional
+import logging
+
+from ap_common import copy_file
+from ap_common.constants import NORMALIZED_HEADER_FILENAME
+
+logger = logging.getLogger(__name__)
+
+
+def check_masters_exist(
+    date_dir: Path,
+    dark: Optional[Dict[str, str]] = None,
+    bias: Optional[Dict[str, str]] = None,
+    flat: Optional[Dict[str, str]] = None,
+) -> Dict[str, bool]:
+    """
+    Check if specific master calibration files already exist in date directory.
+
+    Args:
+        date_dir: Directory to check for existing masters
+        dark: Dark master metadata dict (or None if not needed)
+        bias: Bias master metadata dict (or None if not needed)
+        flat: Flat master metadata dict (or None if not needed)
+
+    Returns:
+        Dict with has_dark, has_bias, has_flat boolean flags indicating
+        whether the specific files already exist
+    """
+    result = {"has_dark": False, "has_bias": False, "has_flat": False}
+
+    if not date_dir.exists():
+        return result
+
+    # Check if specific master files exist by filename
+    if dark:
+        dark_name = Path(dark[NORMALIZED_HEADER_FILENAME]).name
+        result["has_dark"] = (date_dir / dark_name).exists()
+
+    if bias:
+        bias_name = Path(bias[NORMALIZED_HEADER_FILENAME]).name
+        result["has_bias"] = (date_dir / bias_name).exists()
+
+    if flat:
+        flat_name = Path(flat[NORMALIZED_HEADER_FILENAME]).name
+        result["has_flat"] = (date_dir / flat_name).exists()
+
+    return result
+
+
+def copy_master_to_blink(
+    master_metadata: Dict[str, str],
+    dest_dir: Path,
+    dry_run: bool = False,
+) -> bool:
+    """
+    Copy master calibration frame to blink directory.
+
+    Args:
+        master_metadata: Metadata dict for master frame (includes 'filename')
+        dest_dir: Destination directory (DATE directory)
+        dry_run: If True, log action but don't copy
+
+    Returns:
+        True if copied (or would be copied in dry-run), False if skipped
+    """
+    source_path = Path(master_metadata[NORMALIZED_HEADER_FILENAME])
+    dest_path = dest_dir / source_path.name
+
+    if dest_path.exists():
+        logger.debug(f"Master already exists, skipping: {dest_path.name}")
+        return False
+
+    if dry_run:
+        logger.debug(f"[DRY-RUN] Would copy: {source_path.name} -> {dest_dir}")
+        return True
+
+    logger.debug(f"Copying: {source_path.name} -> {dest_dir}")
+    copy_file(str(source_path), str(dest_path))
+    return True
