+++ ap_copy_master_to_blink/statistics.py
@@ -1,0 +1,111 @@
+"""
+Statistics structure definition and formatting
+
+Generated By: Claude Code (Claude Sonnet 4.5)
+"""
+
+from typing import TypedDict
+
+from .config import EXIT_SUCCESS, EXIT_ERROR
+
+
+class Statistics(TypedDict):
+    """Statistics dictionary for process_blink_directory results."""
+
+    frame_count: int
+    target_count: int
+    date_count: int
+    filter_count: int
+    darks_needed: int
+    darks_present: int
+    biases_needed: int
+    biases_present: int
+    flats_needed: int
+    flats_present: int
+    configs_processed: int
+
+
+def create_statistics() -> Statistics:
+    """
+    Create an initialized statistics dictionary.
+
+    Returns:
+        Statistics dictionary with all counters set to 0
+    """
+    return {
+        "frame_count": 0,
+        "target_count": 0,
+        "date_count": 0,
+        "filter_count": 0,
+        "darks_needed": 0,
+        "darks_present": 0,
+        "biases_needed": 0,
+        "biases_present": 0,
+        "flats_needed": 0,
+        "flats_present": 0,
+        "configs_processed": 0,
+    }
+
+
+def print_summary(stats: Statistics) -> None:
+    """
+    Print summary of processing results.
+
+    Args:
+        stats: Statistics dictionary from process_blink_directory
+    """
+
+    def plural(count: int, singular: str) -> str:
+        """Format count with singular/plural form."""
+        return f"{count} {singular}{'s' if count != 1 else ''}"
+
+    def status_indicator(present: int, needed: int) -> str:
+        """Return status indicator based on present vs needed."""
+        return "ok" if present >= needed else "MISSING!"
+
+    print(f"\n{'='*70}")
+    print("Summary")
+    print(f"{'='*70}")
+    print(
+        f"Frames: {stats['frame_count']} lights "
+        f"({plural(stats['target_count'], 'target')}, "
+        f"{plural(stats['date_count'], 'date')}, "
+        f"{plural(stats['filter_count'], 'filter')})"
+    )
+
+    # CRITICAL: Output order MUST be: Biases, Darks, Flats
+    # CRITICAL: Bias MUST ALWAYS be shown (even if 0 of 0)
+    # This order has regressed multiple times - DO NOT CHANGE without updating tests
+    # See test_print_summary_output_order() in tests/test_copy_masters.py
+    print(
+        f"Biases: {stats['biases_present']} of {stats['biases_needed']} | "
+        f"{status_indicator(stats['biases_present'], stats['biases_needed'])}"
+    )
+    print(
+        f"Darks:  {stats['darks_present']} of {stats['darks_needed']} | "
+        f"{status_indicator(stats['darks_present'], stats['darks_needed'])}"
+    )
+    print(
+        f"Flats:  {stats['flats_present']} of {stats['flats_needed']} | "
+        f"{status_indicator(stats['flats_present'], stats['flats_needed'])}"
+    )
+    print(f"{'='*70}\n")
+
+
+def get_exit_code(stats: Statistics) -> int:
+    """
+    Determine exit code based on statistics.
+
+    Args:
+        stats: Statistics dictionary from process_blink_directory
+
+    Returns:
+        EXIT_SUCCESS if all required masters are present, EXIT_ERROR otherwise
+    """
+    darks_missing = stats["darks_needed"] - stats["darks_present"]
+    flats_missing = stats["flats_needed"] - stats["flats_present"]
+
+    if darks_missing > 0 or flats_missing > 0:
+        return EXIT_ERROR
+
+    return EXIT_SUCCESS
