+++ tests/test_file_operations.py
@@ -1,0 +1,219 @@
+"""
+Tests for file_operations.py module.
+
+Generated By: Claude Code (Claude Sonnet 4.5)
+"""
+
+import unittest
+from pathlib import Path
+from unittest.mock import patch
+import tempfile
+
+from ap_copy_master_to_blink.file_operations import (
+    copy_master_to_blink,
+    check_masters_exist,
+)
+from ap_common.constants import NORMALIZED_HEADER_FILENAME
+
+
+class TestFileOperations(unittest.TestCase):
+    """Test file operation functions."""
+
+    @patch("ap_copy_master_to_blink.file_operations.copy_file")
+    def test_copy_master_to_blink_new_file(self, mock_copy_file):
+        """Test copying master when file doesn't exist."""
+        master_metadata = {NORMALIZED_HEADER_FILENAME: "/library/dark.xisf"}
+        dest_dir = Path("/blink/DATE_2024-01-15")
+
+        with patch("pathlib.Path.exists", return_value=False):
+            result = copy_master_to_blink(master_metadata, dest_dir, dry_run=False)
+
+        self.assertTrue(result)
+        mock_copy_file.assert_called_once()
+
+    def test_copy_master_to_blink_existing_file(self):
+        """Test skipping copy when file already exists."""
+        master_metadata = {NORMALIZED_HEADER_FILENAME: "/library/dark.xisf"}
+        dest_dir = Path("/blink/DATE_2024-01-15")
+
+        with patch("pathlib.Path.exists", return_value=True):
+            result = copy_master_to_blink(master_metadata, dest_dir, dry_run=False)
+
+        self.assertFalse(result)
+
+    def test_copy_master_to_blink_dry_run(self):
+        """Test dry-run mode doesn't copy files."""
+        master_metadata = {NORMALIZED_HEADER_FILENAME: "/library/dark.xisf"}
+        dest_dir = Path("/blink/DATE_2024-01-15")
+
+        with patch("pathlib.Path.exists", return_value=False):
+            with patch(
+                "ap_copy_master_to_blink.file_operations.copy_file"
+            ) as mock_copy:
+                result = copy_master_to_blink(master_metadata, dest_dir, dry_run=True)
+
+        self.assertTrue(result)
+        mock_copy.assert_not_called()
+
+    def test_check_masters_exist_no_directory(self):
+        """Test check_masters_exist with non-existent directory."""
+        dark = {NORMALIZED_HEADER_FILENAME: "/library/masterDark.xisf"}
+        bias = {NORMALIZED_HEADER_FILENAME: "/library/masterBias.xisf"}
+        flat = {NORMALIZED_HEADER_FILENAME: "/library/masterFlat.xisf"}
+
+        result = check_masters_exist(Path("/nonexistent"), dark, bias, flat)
+
+        self.assertFalse(result["has_dark"])
+        self.assertFalse(result["has_bias"])
+        self.assertFalse(result["has_flat"])
+
+    def test_check_masters_exist_no_files_requested(self):
+        """Test check_masters_exist when no specific files requested."""
+        with patch("pathlib.Path.exists", return_value=True):
+            result = check_masters_exist(Path("/blink/DATE_2024-01-15"))
+
+        self.assertFalse(result["has_dark"])
+        self.assertFalse(result["has_bias"])
+        self.assertFalse(result["has_flat"])
+
+    def test_check_masters_exist_dark_exists(self):
+        """Test check_masters_exist finds existing dark file."""
+        dark = {NORMALIZED_HEADER_FILENAME: "/library/masterDark.xisf"}
+
+        def mock_exists(self):
+            # Date directory exists, dark file exists
+            posix_path = (
+                self.as_posix()
+                if hasattr(self, "as_posix")
+                else str(self).replace("\\", "/")
+            )
+            return posix_path in [
+                "/blink/DATE_2024-01-15",
+                "/blink/DATE_2024-01-15/masterDark.xisf",
+            ]
+
+        with patch("pathlib.Path.exists", mock_exists):
+            result = check_masters_exist(Path("/blink/DATE_2024-01-15"), dark=dark)
+
+        self.assertTrue(result["has_dark"])
+        self.assertFalse(result["has_bias"])
+        self.assertFalse(result["has_flat"])
+
+    def test_check_masters_exist_flat_exists(self):
+        """Test check_masters_exist finds existing flat file."""
+        flat = {NORMALIZED_HEADER_FILENAME: "/library/masterFlat.xisf"}
+
+        def mock_exists(self):
+            # Date directory exists, flat file exists
+            posix_path = (
+                self.as_posix()
+                if hasattr(self, "as_posix")
+                else str(self).replace("\\", "/")
+            )
+            return posix_path in [
+                "/blink/DATE_2024-01-15",
+                "/blink/DATE_2024-01-15/masterFlat.xisf",
+            ]
+
+        with patch("pathlib.Path.exists", mock_exists):
+            result = check_masters_exist(Path("/blink/DATE_2024-01-15"), flat=flat)
+
+        self.assertFalse(result["has_dark"])
+        self.assertFalse(result["has_bias"])
+        self.assertTrue(result["has_flat"])
+
+    def test_check_masters_exist_all_exist(self):
+        """Test check_masters_exist when all requested files exist."""
+        dark = {NORMALIZED_HEADER_FILENAME: "/library/masterDark.xisf"}
+        bias = {NORMALIZED_HEADER_FILENAME: "/library/masterBias.xisf"}
+        flat = {NORMALIZED_HEADER_FILENAME: "/library/masterFlat.xisf"}
+
+        def mock_exists(self):
+            # Date directory and all master files exist
+            posix_path = (
+                self.as_posix()
+                if hasattr(self, "as_posix")
+                else str(self).replace("\\", "/")
+            )
+            return posix_path in [
+                "/blink/DATE_2024-01-15",
+                "/blink/DATE_2024-01-15/masterDark.xisf",
+                "/blink/DATE_2024-01-15/masterBias.xisf",
+                "/blink/DATE_2024-01-15/masterFlat.xisf",
+            ]
+
+        with patch("pathlib.Path.exists", mock_exists):
+            result = check_masters_exist(
+                Path("/blink/DATE_2024-01-15"), dark, bias, flat
+            )
+
+        self.assertTrue(result["has_dark"])
+        self.assertTrue(result["has_bias"])
+        self.assertTrue(result["has_flat"])
+
+    def test_check_masters_exist_partial(self):
+        """Test check_masters_exist when only some files exist."""
+        dark = {NORMALIZED_HEADER_FILENAME: "/library/masterDark.xisf"}
+        bias = {NORMALIZED_HEADER_FILENAME: "/library/masterBias.xisf"}
+        flat = {NORMALIZED_HEADER_FILENAME: "/library/masterFlat.xisf"}
+
+        def mock_exists(self):
+            # Date directory exists, only dark exists
+            posix_path = (
+                self.as_posix()
+                if hasattr(self, "as_posix")
+                else str(self).replace("\\", "/")
+            )
+            return posix_path in [
+                "/blink/DATE_2024-01-15",
+                "/blink/DATE_2024-01-15/masterDark.xisf",
+            ]
+
+        with patch("pathlib.Path.exists", mock_exists):
+            result = check_masters_exist(
+                Path("/blink/DATE_2024-01-15"), dark, bias, flat
+            )
+
+        self.assertTrue(result["has_dark"])
+        self.assertFalse(result["has_bias"])
+        self.assertFalse(result["has_flat"])
+
+    def test_copy_file_receives_strings_not_paths(self):
+        """Test that copy_file is called with strings, not Path objects.
+
+        Regression test for bug where Path objects were passed to copy_file
+        which expects strings.
+        """
+        from ap_copy_master_to_blink.file_operations import copy_master_to_blink
+
+        with tempfile.TemporaryDirectory() as tmpdir:
+            # Create source file
+            library_dir = Path(tmpdir) / "library"
+            library_dir.mkdir()
+            source_file = library_dir / "masterDark.xisf"
+            source_file.write_text("fake xisf")
+
+            # Create dest directory
+            dest_dir = Path(tmpdir) / "blink" / "DATE_2024-01-15"
+            dest_dir.mkdir(parents=True)
+
+            master_metadata = {NORMALIZED_HEADER_FILENAME: str(source_file)}
+
+            # Should not raise AttributeError about
+            # Path object missing 'split'
+            try:
+                result = copy_master_to_blink(master_metadata, dest_dir, dry_run=False)
+                self.assertTrue(result)
+                # Verify file was copied
+                dest_file = dest_dir / source_file.name
+                self.assertTrue(dest_file.exists())
+            except AttributeError as e:
+                if "'WindowsPath' object has no attribute 'split'" in str(
+                    e
+                ) or "'PosixPath' object has no attribute 'split'" in str(e):
+                    self.fail(f"copy_file received Path object instead of string: {e}")
+                raise
+
+
+if __name__ == "__main__":
+    unittest.main()
