+++ tests/test_main.py
@@ -6,9 +6,16 @@ Generated By: Claude Code (Claude Sonnet 4.5)
 
 import unittest
 from io import StringIO
+from pathlib import Path
 from unittest.mock import patch
 
-from ap_copy_master_to_blink.__main__ import EXIT_ERROR, EXIT_SUCCESS, main
+from ap_copy_master_to_blink.__main__ import (
+    EXIT_ERROR,
+    EXIT_SUCCESS,
+    main,
+    validate_directories,
+    print_header,
+)
 
 
 class TestMainCLI(unittest.TestCase):
@@ -275,6 +282,67 @@ class TestMainCLI(unittest.TestCase):
         result = main()
 
         self.assertEqual(result, EXIT_ERROR)
+
+
+class TestCLIFunctions(unittest.TestCase):
+    """Test CLI helper functions."""
+
+    def test_validate_directories_both_valid(self):
+        """Test validation when both directories are valid."""
+        with patch("pathlib.Path.exists", return_value=True):
+            with patch("pathlib.Path.is_dir", return_value=True):
+                is_valid, error = validate_directories(Path("/library"), Path("/blink"))
+
+                self.assertTrue(is_valid)
+                self.assertIsNone(error)
+
+    def test_validate_directories_library_not_exists(self):
+        """Test validation when library directory doesn't exist."""
+        with patch("pathlib.Path.exists", side_effect=[False, True]):
+            is_valid, error = validate_directories(Path("/library"), Path("/blink"))
+
+            self.assertFalse(is_valid)
+            self.assertIn("Library directory does not exist", error)
+
+    def test_validate_directories_library_not_dir(self):
+        """Test validation when library path is not a directory."""
+        with patch("pathlib.Path.exists", return_value=True):
+            with patch("pathlib.Path.is_dir", side_effect=[False, True]):
+                is_valid, error = validate_directories(Path("/library"), Path("/blink"))
+
+                self.assertFalse(is_valid)
+                self.assertIn("Library path is not a directory", error)
+
+    def test_validate_directories_blink_not_exists(self):
+        """Test validation when blink directory doesn't exist."""
+        with patch("pathlib.Path.exists", side_effect=[True, False]):
+            with patch("pathlib.Path.is_dir", side_effect=[True]):
+                is_valid, error = validate_directories(Path("/library"), Path("/blink"))
+
+                self.assertFalse(is_valid)
+                self.assertIn("Blink directory does not exist", error)
+
+    def test_validate_directories_blink_not_dir(self):
+        """Test validation when blink path is not a directory."""
+        with patch("pathlib.Path.exists", return_value=True):
+            with patch("pathlib.Path.is_dir", side_effect=[True, False]):
+                is_valid, error = validate_directories(Path("/library"), Path("/blink"))
+
+                self.assertFalse(is_valid)
+                self.assertIn("Blink path is not a directory", error)
+
+    @patch("sys.stdout", new_callable=StringIO)
+    def test_print_header(self, mock_stdout):
+        """Test header printing."""
+        library_path = Path("/library")
+        blink_path = Path("/blink")
+        print_header(library_path, blink_path, dry_run=True)
+
+        output = mock_stdout.getvalue()
+        self.assertIn("ap-copy-master-to-blink", output)
+        self.assertIn(f"Library: {library_path}", output)
+        self.assertIn(f"Blink:   {blink_path}", output)
+        self.assertIn("Dry-run: True", output)
 
 
 if __name__ == "__main__":
