+++ tests/test_path_utils.py
@@ -1,0 +1,83 @@
+"""
+Tests for path_utils.py module.
+
+Generated By: Claude Code (Claude Sonnet 4.5)
+"""
+
+import unittest
+from pathlib import Path
+
+from ap_copy_master_to_blink.path_utils import get_date_directory
+
+
+class TestPathUtils(unittest.TestCase):
+    """Test path utility functions."""
+
+    def test_get_date_directory_from_filter_dir(self):
+        """Test extracting DATE directory from FILTER directory."""
+        lights_dir = Path("/blink/M31/DATE_2024-01-15/FILTER_Ha")
+        result = get_date_directory(lights_dir, date_dir_pattern=r"^DATE_.*")
+
+        self.assertEqual(result.name, "DATE_2024-01-15")
+
+    def test_get_date_directory_from_date_dir(self):
+        """Test when already in DATE directory."""
+        lights_dir = Path("/blink/M31/DATE_2024-01-15")
+        result = get_date_directory(lights_dir, date_dir_pattern=r"^DATE_.*")
+
+        self.assertEqual(result.name, "DATE_2024-01-15")
+
+    def test_get_date_directory_search_upward(self):
+        """Test searching upward for DATE directory."""
+        lights_dir = Path("/blink/M31/DATE_2024-01-15/FILTER_Ha/subdir")
+        result = get_date_directory(lights_dir, date_dir_pattern=r"^DATE_.*")
+
+        self.assertEqual(result.name, "DATE_2024-01-15")
+
+    def test_get_date_directory_custom_pattern(self):
+        """Test with custom date directory pattern."""
+        # Use NIGHT_ prefix instead of DATE_
+        lights_dir = Path("/blink/M31/NIGHT_2024-01-15/FILTER_Ha")
+        result = get_date_directory(lights_dir, date_dir_pattern=r"^NIGHT_.*")
+
+        self.assertEqual(result.name, "NIGHT_2024-01-15")
+
+    def test_get_date_directory_pattern_mismatch_returns_leaf(self):
+        """Test fallback to leaf directory when pattern doesn't match."""
+        lights_dir = Path("/blink/M31/OBS_2024-01-15/FILTER_Ha")
+        # Pattern expects DATE_ but directory has OBS_
+        result = get_date_directory(lights_dir, date_dir_pattern=r"^DATE_.*")
+
+        # Should return leaf directory (FILTER_Ha) since no match found
+        self.assertEqual(result, lights_dir)
+
+    def test_get_date_directory_no_hardcoded_filter_check(self):
+        """Test that FILTER_ prefix is not special - uses pattern only."""
+        # Directory named FILTER_X but pattern looks for FILTER_
+        lights_dir = Path("/blink/M31/FILTER_Ha/subdir")
+        result = get_date_directory(lights_dir, date_dir_pattern=r"^FILTER_.*")
+
+        # Should match FILTER_Ha since pattern matches it
+        self.assertEqual(result.name, "FILTER_Ha")
+
+    def test_get_date_directory_multiple_pattern_matches(self):
+        """Test that first match (closest to leaf) is returned."""
+        # Both DATE_2024-01-15 and DATE_2024-01-14 match pattern
+        lights_dir = Path("/blink/DATE_2024-01-14/TARGET_M31/DATE_2024-01-15/FILTER_Ha")
+        result = get_date_directory(lights_dir, date_dir_pattern=r"^DATE_.*")
+
+        # Should return the closest match (DATE_2024-01-15)
+        self.assertEqual(result.name, "DATE_2024-01-15")
+
+    def test_get_date_directory_at_filesystem_root(self):
+        """Test behavior when reaching filesystem root without match."""
+        # Only root directory, no matching directories
+        lights_dir = Path("/lights")
+        result = get_date_directory(lights_dir, date_dir_pattern=r"^DATE_.*")
+
+        # Should return original directory (leaf)
+        self.assertEqual(result, lights_dir)
+
+
+if __name__ == "__main__":
+    unittest.main()
